requirements.txt
flask
opencv-python
ultralytics
python-dotenv
google-generativeai
requests
Pillow